move mapping test
